+++ frontend/src/components/pages/signin/index.jsx
@@ -1,4 +1,4 @@
-import { useState, useContext } from 'react'
+import { useContext } from 'react'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import Cookies from 'js-cookie'
@@ -9,9 +9,7 @@ import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { userSchema } from 'src/validation/auth'
 
-import { Button } from 'src/components/shared/Button'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faAsterisk, faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons'
+import { SignupSigninForm } from 'src/components/shared/SigninSignupForm'
 
 export const SignInPage = () => {
   const router = useRouter()
@@ -54,78 +52,21 @@ export const SignInPage = () => {
     }
   }
 
-  // パスワード表示/非表示切り替え
-  const [isRevealPassword, setIsRevealPassword] = useState(false)
-  const togglePassword = () => {
-    setIsRevealPassword((prevState) => !prevState)
-  }
-
   return (
     <div className="my-40 grid place-content-center place-items-center">
       <h1 className="text-center text-2xl font-bold tracking-widest text-dark-blue">ログイン</h1>
-      <form
-        noValidate
-        onSubmit={handleSubmit(onSubmit)}
-        className="grid place-content-center place-items-center"
-      >
-        <div className="form-control my-6 h-32 w-96">
-          <label htmlFor="email" className="label">
-            <span className="text-base text-dark-black">メールアドレス</span>
-          </label>
-          <input
-            id="email"
-            type="email"
-            autoComplete="email"
-            {...register('email')}
-            placeholder="your@email.com"
-            className="w-ful input input-bordered input-secondary input-md border-dark-pink bg-ligth-white text-base text-dark-black"
-          />
-          {errors.email && <p className="label text-base text-dark-pink">{errors.email.message}</p>}
-        </div>
-        <div className="form-control mb-12 h-32 w-96">
-          <label htmlFor="password" className="label">
-            <span className="text-base text-dark-black">パスワード</span>
-            <span className="label-text-alt text-sm text-dark-black">
-              <FontAwesomeIcon icon={faAsterisk} className="mr-1 text-xs text-dark-pink" />
-              6文字以上の半角英数字
-            </span>
-          </label>
-          <div className="flex items-center">
-            <div className="relative">
-              <input
-                id="password"
-                type={isRevealPassword ? 'text' : 'password'}
-                autoComplete="current-password"
-                {...register('password')}
-                placeholder="password"
-                className="input input-bordered input-secondary input-md w-96 border-dark-pink bg-ligth-white text-base text-dark-black"
-              />
-              <span onClick={togglePassword} role="presentation" className="absolute right-3 top-3">
-                {isRevealPassword ? (
-                  <FontAwesomeIcon icon={faEye} />
-                ) : (
-                  <FontAwesomeIcon icon={faEyeSlash} />
-                )}
-              </span>
-              {errors.password && (
-                <p className="label text-base text-dark-pink">{errors.password.message}</p>
-              )}
-            </div>
-          </div>
-        </div>
-        <Button
-          btnType="submit"
-          disabled={!dirtyFields.email || !dirtyFields.password}
-          addStyle="btn-secondary h-16 w-40"
-        >
-          ログイン
-        </Button>
-      </form>
-      <p className="mb-6 mt-28 text-base text-dark-black">新規登録はこちら</p>
-      <Link href="/signup" passHref>
-        <Button btnType="button" addStyle="btn-primary h-16 w-40">
-          新規登録
-        </Button>
+
+      <SignupSigninForm
+        register={register}
+        handleSubmit={handleSubmit}
+        dirtyFields={dirtyFields}
+        errors={errors}
+        onSubmit={onSubmit}
+        buttonName="ログイン"
+        addStyle="btn-secondary h-16 w-40"
+      />
+      <Link href="/signup" passHref className="link-hover link my-2">
+        新規登録はこちら
       </Link>
     </div>
   )
